pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nl2sql"
version = "0.1.0"
description = "Natural Language to SQL"
readme = "README.rst"
authors = [
  {name = "Yeshwanth Reddy", email = "yeshwanth@divami.com"}
]
maintainers = [
  {name = "Yeshwanth Reddy", email = "yeshwanth@divami.com"}
]
classifiers = [

]
license = {text = "MIT license"}
dependencies = [
  "typer",
  "nbdev",
  "mkdocs-material>=9.6.18",
  "mkdocs-mermaid2-plugin>=1.2.1",
  "mkdocs-awesome-pages-plugin>=2.10.1",
  "mkdocs-minify-plugin>=0.8.0",
  "mkdocs-git-revision-date-localized-plugin>=1.4.7",
  "pymdown-extensions>=10.16.1",
  "sqlalchemy>=2.0.43",
  "torch-snippets>=0.558",
  "duckdb>=1.3.2",
  "httpx>=0.24.0",
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "python-dotenv>=1.0.0",
  "pydantic-ai>=0.0.14",
  "google-generativeai>=0.8.0",
  "annotated-types>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "coverage",  # testing
    "mypy",  # linting
    "pytest",  # testing
    "ruff"  # linting
]

[project.urls]

bugs = "https://github.com/yeshwanth-divami/nl2sql/issues"
changelog = "https://github.com/yeshwanth-divami/nl2sql/blob/master/changelog.md"
homepage = "https://github.com/yeshwanth-divami/nl2sql"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.*"]

[project.scripts]
nl2sql = "nl2sql:cli"

# Mypy
# ----

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

